refactor(dhcp): split lease-file parsing out of the dhcp god-file (WS-D)
internal/dhcp/dhcp.go was a 1002-line god-file holding two distinct concerns:
the live DHCP transaction Monitor (capture + phase timing) and the OS lease-file
readers. Split the lease concern into its own file (pure move — no API, signature
or logic change; verified by an identical top-level declaration set and an
identical sorted code-line multiset against the original):

- dhcp.go (524)  — Phase/Timing/Transaction/Monitor + capture + phase timing
- lease.go (486) — LeaseInfo + GetLeaseInfo and the darwin/linux/NM/networkd
                   lease-file parsers

Both files are now well under the size threshold, removing dhcp.go as a STRICT
file-size-gate blocker for WS-Z. Imports split per file via goimports; the
package doc comment stays attached to dhcp.go only.

Advances the Architecture Completion Plan WS-D (god-file decomposition).
